requirements.txt
pynacl==1.2.1
